restructure code for frontend
