+++ frontend/src/pages/Main.jsx
@@ -2,7 +2,6 @@ import React from "react";
 import { Outlet } from "react-router-dom";
 import Header from "../components/layout/Header";
 import Sidebar from "../components/layout/Sidebar";
-import './styles/admin.css'
 
 const Main = (props) => {
   return (
